requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
